requirements.txt
Pillow==7.1.2
pypiwin32==223
pywin32==228
